+++ NEWS.md
@@ -1,7 +1,7 @@
 # NWCTrends
 
 # Version 1.25
-This is the version used in the 2020 PNW Salmonid Status Review. 
+This is the version used in the 2020 NWFSC PNW Salmonid Viability Report. 
 
 Add more meta data: Method (Survey, Model), Citation, and Contributor. A `metadata.csv` file is added to each Output folder.
 
@@ -11,6 +11,6 @@ Add more meta data: Method (Survey, Model), Citation, and Contributor. A `metada
 * Added a `NEWS.md` file
 
 # Version 1.0
-This is the version used in the 2015 PNW Salmonid Status Review. 
+This is the version used in the 2015 NWFSC PNW Salmonid Viability Report.