+++ NEWS.md
@@ -1,5 +1,15 @@
 # NWCTrends
 
+# Version 1.26
+
+Added a data object `nwctrends.pallete` with red, blue, black, green and white defined (in a list). To change the colors, the user runs a line like
+```
+nwctrends.palette$red <- "yellow"
+```
+to say change red to yellow.
+
+Added `show.all.fracwild` to `NWCTrends_report()` so that that can be changed from the main function.
+
 # Version 1.25
 This is the version used in the 2020 NWFSC PNW Salmonid Viability Report.